Implement MVVM Architecture
