+++ Shared/Features/Freaks/View/FreaksView.swift
@@ -1,0 +1,19 @@
+import SwiftUI
+
+struct FreaksView: View {
+    var viewModel = FreaksViewModel()
+
+    var body: some View {
+        NavigationView {
+            List(viewModel.freaks, id: \.firstName) { freak in
+                ListItemView(freak: freak)
+            }
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        FreaksView()
+    }
+}
