+++ Shared/Features/Freaks/ViewModel/FreaksViewModel.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+final class FreaksViewModel {
+    var freaks: [Freak] = [
+        Freak(imageName: "person.crop.square", firstName: "Vlad", lastName: "Stanescu"),
+        Freak(imageName: "person.crop.square", firstName: "Mihai", lastName: "Pantiru"),
+        Freak(imageName: "person.crop.square", firstName: "Raluca", lastName: "Ionescu"),
+    ]
+}
